+++ src/dummy_write_handler.hpp
@@ -15,8 +15,8 @@ public:
   DummyWriteHandler(DummyWriteHandler &&) = default;
   DummyWriteHandler &operator=(DummyWriteHandler &&) = default;
 
-  int handle(std::span<std::byte const> buf,
-             __off64_t offset [[maybe_unused]]) noexcept override {
+  ssize_t handle(std::span<std::byte const> buf,
+                 __off64_t offset [[maybe_unused]]) noexcept override {
     return buf.size();
   }
 };
